# src/models/Character.ts
export interface Character {
    id: string;
    nombre: string;
    altura: string;
    masa:string;
    color_pelo: string;
    color_ojos: string;
    color_piel: string;
    nacimiento: string;
    genero: string;
}