# src/swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "API para manejar personajes de Star Wars",
      "version": "1.0.0",
      "title": "Star Wars API"
    },
    "host": "",
    "schemes": [
        ""
    ],
    "basePath": "/dev",
    "paths": {
      "/characters": {
        "get": {
          "summary": "Obtener todos los personajes",
          "produces": ["application/json"],
          "responses": {
            "200": {
              "description": "Lista de personajes"
            },
            "500": {
              "description": "Error al obtener personajes"
            }
          },
          "deprecated": false,
          "x-cors": {
            "allowOrigins": ["*"],
            "allowMethods": ["GET", "OPTIONS"],
            "allowHeaders": ["Content-Type", "X-Requested-With", "Authorization"]
          }
        },
        "post": {
          "summary": "Crear un nuevo personaje",
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "description": "Datos del personaje",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Character"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Personaje creado exitosamente"
            },
            "500": {
              "description": "Error al crear el personaje"
            }
          },
          "deprecated": false,
          "x-cors": {
            "allowOrigins": ["*"],
            "allowMethods": ["POST", "OPTIONS"],
            "allowHeaders": ["Content-Type", "X-Requested-With", "Authorization"]
          }
        }
      }
    },
    "definitions": {
      "Character": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "height": {
            "type": "number"
          },
          "mass": {
            "type": "string"
          },
          "hair_color": {
            "type": "string"
          },
          "skin_color": {
            "type": "string"
          },
          "eye_color": {
            "type": "string"
          },
          "birth_year": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          }
        },
        "required": ["name", "height", "mass", "hair_color", "skin_color", "eye_color", "birth_year", "gender"]
      }
    }
  }
  